Correção do backEnd impossibilitar deletar usuário com mesmo Id

+++ codigo-fonte/time-sheet-api/TimeSheet/Commands/DeleteUserCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿namespace TimeSheet.Commands {
+﻿using System.Security.Claims;
+
+namespace TimeSheet.Commands {
     public class DeleteUserCommandHandler
         : ICommandHandler<DeleteUserCommand, DeleteUserCommandResult> {
 
@@ -11,6 +13,7 @@
 
             var user = await _repository
                 .FindUser(command.UserId);
+            var currenteid = Guid.Parse(ClaimsPrincipal.Current.Claims.First(c => c.Type == ClaimTypes.Sid).Value);
 
             if (user is null) {
                 return new DeleteUserCommandResult {
@@ -18,8 +21,16 @@
                     Status = DeleteUserCommandResultStatus.UserNotFound
                 };
             }
+            if (currenteid == user.Id)
+            {
+                return new DeleteUserCommandResult
+                {
+                    Message = "Usuário não pode deletar o própio usuário",
+                    Status = DeleteUserCommandResultStatus.UserNotDeleted
+                };
+            }
 
-            await _repository.DeleteUser(user);
+                await _repository.DeleteUser(user);
 
             return new DeleteUserCommandResult {
                 Message = "Usuário deletado.",
